requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Modals==0.5.1
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.0.6
wtforms==3.1.2
zipp==3.20.2
